+++ src/frontend/app/shared-tdoc/components/tdoc-assignform/tdoc-assignform.component.spec.ts
@@ -14,7 +14,7 @@ import {RouterStub} from '@dps/mycms-frontend-commons/dist/angular-commons/testi
 import {TourDocContentUtils} from '../../services/tdoc-contentutils.service';
 import {TourDocAssignFormComponent} from './tdoc-assignform.component';
 import {FormBuilder} from '@angular/forms';
-import {TourDocDataStore, TourDocTeamFilterConfig} from '../../../../../shared/tdoc-commons/services/tdoc-data.store';
+import {TourDocDataStore, TourDocTeamFilterConfig} from '../../../../shared/tdoc-commons/services/tdoc-data.store';
 import {SearchParameterUtils} from '@dps/mycms-commons/dist/search-commons/services/searchparameter.utils';
 import {TourDocDataService} from '../../../../shared/tdoc-commons/services/tdoc-data.service';
 import {SearchFormUtils} from '@dps/mycms-frontend-commons/dist/angular-commons/services/searchform-utils.service';